+++ packages/buy-button/src/__tests__/BuyButton.test.js
@@ -1,16 +1,38 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import BuyButton from '../index'
+import { BuyButton } from '../index'
 
 describe('<BuyButton />', () => {
-  it('should pass', () => {
-    expect(1).toBe(1)
+  let wrapper
+
+  const props = {
+    quantity: 1,
+    skuId: 1,
+    seller: '1',
+    salesChannel: '1',
+    data: { orderForm: { orderFormId: '12321' } },
+    mutate: jest.fn(),
+    afterClick: jest.fn(),
+  }
+
+  beforeEach(() => {
+    wrapper = shallow(<BuyButton {...props}> teste </BuyButton>)
+  })
+
+  it('should match snapshot', () => {
+    expect(wrapper).toMatchSnapshot()
   })
 
   it('should be rendered', () => {
-    const wrapper = shallow(<BuyButton />)
-
     expect(wrapper).toBeDefined()
+  })
+
+  it('should call afterClick and mutate', () => {
+    expect(props.afterClick.mock.calls.length).toBe(0)
+    expect(props.mutate.mock.calls.length).toBe(0)
+    wrapper.find('Button').simulate('click')
+    expect(props.afterClick.mock.calls.length).toBe(1)
+    expect(props.mutate.mock.calls.length).toBe(1)
   })
 })