+++ packages/buy-button/src/index.js
@@ -1,28 +1,29 @@
-import React from 'react'
+import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { graphql } from 'react-apollo'
+import { graphql, compose } from 'react-apollo'
 
 import Button from '@vtex/styleguide/lib/Button'
 
 import addToCartMutation from './mutations/addToCartMutation.gql'
+import orderFormQuery from './queries/orderFormQuery.gql'
 
 /**
  * BuyButton Component. Adds a list of items to the cart.
  */
-class BuyButton extends React.Component {
+export class BuyButton extends Component {
   handleAddToCart = () => {
     const {
+      data: { orderForm: { orderFormId } },
       mutate,
       quantity,
       seller,
       skuId,
       afterClick,
-      orderFormId,
     } = this.props
 
     mutate({
       variables: {
-        orderFormId: orderFormId,
+        orderFormId,
         items: [
           {
             id: parseInt(skuId),
@@ -32,6 +33,7 @@ class BuyButton extends React.Component {
           },
         ],
       },
+      refetchQueries: [{ query: orderFormQuery }],
     })
     afterClick()
   }
@@ -60,8 +62,18 @@ BuyButton.propTypes = {
   mutate: PropTypes.func.isRequired,
   /** Function that will be called after the mutation */
   afterClick: PropTypes.func.isRequired,
-  /** The user's cart id */
-  orderFormId: PropTypes.string.isRequired,
+  /** Property that contains OrderForm response */
+  data: PropTypes.shape({
+    /** Order form used in the buy button */
+    orderForm: PropTypes.shape({
+      /** User's cart id */
+      orderFormId: PropTypes.string.isRequired,
+    }),
+  }),
 }
-
-export default graphql(addToCartMutation)(BuyButton)
+export default compose(
+  graphql(orderFormQuery, {
+    options: { ssr: false },
+  }),
+  graphql(addToCartMutation)
+)(BuyButton)