+++ packages/buy-button/CHANGELOG.md
@@ -7,19 +7,32 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [0.1.5] - 2018-26-04
+## [0.1.16] - 2018-05-07
+### Fixed 
+- Fix buy-button to get orderFormId 
 
-### Bugfix
+## [0.1.9] - 2018-05-07
+
+## [0.1.8] - 2018-05-03
+
+### Added
+
+* **Package Config** Prepublish to package.json.
+* **Tests** Component Tests
+
+## [0.1.5] - 2018-04-26
+
+### Fixed
 
 * **Node Module** Retrieve the version 0.1.1.
 
-## [0.1.2] - 2018-20-04
+## [0.1.2] - 2018-04-20
 
 ### Deprecated
 
 * **Node Module** published a wrong version of the module.
 
-## [0.1.1] - 2018-19-04
+## [0.1.1] - 2018-04-19
 
 ### Added